requirements.txt
python-telegram-bot
firebase-admin
flask
python-dotenv  # <-- Tambahkan ini